feat: 전역 Type 위치 변경, pagenationResponse 타입 추가
