+++ front_end/src/stores/auth.ts
@@ -1,4 +1,4 @@
-import type { AuthToken, User } from '@/services/api';
+import type { AuthToken, User } from '@/types/type';
 import { defineStore } from 'pinia';
 
 interface AuthState {
